requirements.txt
pybit==5.8.0
python-telegram-bot[job-queue]==21.7
